refactor(LC032): split bulk-update analysis helpers
- separate query-source, assignment, statement, and capability analysis
- keep analyzer behavior unchanged while reducing local complexity
